rename package from model to models
